requirements.txt
absl-py==1.4.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
flatbuffers==23.1.21
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
idna==3.4
ipykernel==6.21.2
ipython==8.10.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
keras==2.11.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.0.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.14.1
streamlit~=1.18.1
PyYAML~=6.0